+++ src/app/Scenes/Home/Components/MeetYourNewAdvisorRail.tsx
@@ -1,6 +1,6 @@
 import { ActionType, ContextModule, OwnerType } from "@artsy/cohesion"
 import { EmbeddedCarousel } from "app/Components/EmbeddedCarousel"
-import { HomeFeedModalCarousel } from "app/Scenes/Home/Components/HomeFeedModalCarousel/HomeFeedModalCarousel"
+import { HomeFeedModalCarousel } from "app/Scenes/Home/Components/HomeFeedModalCarousel/ModalCarousel"
 import { navigate, switchTab } from "app/system/navigation/navigate"
 import { useState } from "react"
 import { useTracking } from "react-tracking"
